requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.3
django-heroku==0.0.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==20.9
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.5
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
